feat: endpoints directory
