+++ package.json
@@ -1,6 +1,6 @@
 {
     "name": "@eggermarc/better-auth-usage",
-    "version": "0.1.10",
+    "version": "0.1.11",
     "repository": {
         "type": "git",
         "url": "git+https://github.com/EggerMarc/better-auth-usage.git"
