+++ package/client.ts
@@ -6,8 +6,9 @@ export const usageClient = () => {
         id: "@eggermarc/usage-client",
         $InferServerPlugin: {} as ReturnType<typeof usage>,
         pathMethods: {
-            "/usage/feature": "GET",
-            "/usage/register-customer": "POST",
+            "/usage/features": "GET",
+            "/usage/features/:featureKey": "GET",
+            "/usage/upsert-customer": "POST",
             "/usage/consume": "POST",
             "/usage/check": "GET",
             "/usage/sync": "POST",
