+++ package/endpoints/upsert-customer.ts
@@ -1,0 +1,42 @@
+import { createAuthEndpoint } from "better-auth/api";
+import { getUsageAdapter } from "package/adapter";
+import { customerSchema } from "package/schema";
+
+export function getUpsertCustomerEndpoint() {
+    return createAuthEndpoint("/usage/upsert-customer", {
+        method: "POST",
+        body: customerSchema,
+        middleware: [middleware],
+        metadata: {
+            openapi: {
+                description: "Upserts a customer to the customer table",
+                requestBody: {
+                    required: true,
+                    content: {
+                        "application/json": {
+                            schema: {
+                                type: "object",
+                                properties: {
+                                    referenceId: { type: "string" },
+                                    featureType: { type: "string" },
+                                    name: { type: "string" },
+                                    email: { type: "string" }
+                                },
+                                required: ["referenceId", "featureType"],
+                            },
+                        },
+                    },
+                },
+                responses: {
+                    200: {
+                        description: "Successful Upsert",
+                    },
+                },
+            },
+        },
+    }, async (ctx) => {
+        const adapter = getUsageAdapter(ctx.context);
+        const customer = await adapter.upsertCustomer(ctx.body);
+        return customer
+    })
+}
